Move sync logic from cmd/sync to libs/sync
