--- cmd/sync/sync_test.go
@@ -1,12 +1,0 @@
-package sync
-
-import (
-	"testing"
-)
-
-func TestItSyncs(t *testing.T) {
-	// ctx := context.Background()
-	// root.RootCmd.SetArgs([]string{"sync"})
-	// err := root.RootCmd.ExecuteContext(ctx)
-	// assert.NoError(t, err)
-}
