+++ libs/sync/snapshot.go
